maint: add dhat to do some heap/memory profiling, improve memory usage
This adds a new feature `dhat` that when turned on will enable dhat
allocator. It should not be used when producing released builds for
distributions but will need to be enabled with the release profile
builds for more accurate heap/memory representation, as opposed to debug
builds which will likely differ greatly from what would be seen in
production.

A few changes are made to improve memory usage:
1. The bytes initially read into `DicomElement` are moved into the
   `PixelDataSliceInfo` rather than copied.
2. When loading `PixelDataSliceInfo` into `PixelDataSlice` variants the
   bytes are also moved rather than copied.
3. Removed the `WindowLevel` being computed for every call to
   `get_pixel()` to instead be computed once when creating the iterator.